refactor: move open ai chat completions to its own folder
this will allow us to create a folder for responses api
