+++ lib/llm_gateway.rb
@@ -13,8 +13,9 @@ require_relative "llm_gateway/adapters/claude/client"
 require_relative "llm_gateway/adapters/claude/input_mapper"
 require_relative "llm_gateway/adapters/claude/output_mapper"
 require_relative "llm_gateway/adapters/open_ai/client"
-require_relative "llm_gateway/adapters/open_ai/input_mapper"
-require_relative "llm_gateway/adapters/open_ai/output_mapper"
+require_relative "llm_gateway/adapters/open_ai/file_output_mapper"
+require_relative "llm_gateway/adapters/open_ai/chat_completions/input_mapper"
+require_relative "llm_gateway/adapters/open_ai/chat_completions/output_mapper"
 require_relative "llm_gateway/adapters/groq/client"
 require_relative "llm_gateway/adapters/groq/input_mapper"
 require_relative "llm_gateway/adapters/groq/output_mapper"
